requirements.txt
auto-preprocess
basix==0.0.4
category-encoders==2.5.0
click==8.0.4
dill==0.3.5.1
fastparquet==0.8.1
GitPython==3.1.27
ipykernel==6.9.1 
matplotlib==3.5.2
pandas==1.4.2
lightgbm==3.3.2
psycopg2-binary==2.9.3
pyarrow==8.0.0
python-dotenv==0.20.0
PyYAML==6.0
optbinning==0.15.0
requests==2.27.1
scikit-learn==0.23.2
scikit-optimize==0.9.0
shap==0.41.0
sklearn-pandas==1.5.0
skorecard==1.6.4
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sqlalchemy==1.1.2
statsmodels==0.12.2
tqdm==4.64
unidecode==1.3.4
xgboost==1.6.1